pyproject.toml
[project]
name = "hewe-nlp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "langchain>=0.3.27",
    "langchain-ollama>=0.3.6",
    "langchain-openai>=0.3.30",
    "pillow>=11.3.0",
    "pymysql>=1.1.2",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.1",
]
